Added blin-phong support

+++ Ascension Core/Timer.cpp
@@ -58,5 +58,5 @@ double Timer::GetFrameTime()
 	if (tickCount < 0.0f)
 		tickCount = 0;
 
-	return float(tickCount) / CountsPerSecond;
+	return double(tickCount) / CountsPerSecond;
 }
